requirements.txt
schedule
ffmpeg-python
screen_recorder_sdk
pyaudio